pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dca"
version = "1.3.0"
description = "A Python implementation of mean-field Direct Coupling Analysis, adapted from the dca.m script from http://dca.rice.edu/portal/dca/."
authors = [
    { name = "Jonathan Martin", email = "jonathan.martin3@utdallas.edu" },
]
license = { text = "MIT" }
readme = "README.md"
keywords = ["DCA", "Machine Learning", "Potts", "Proteins"]
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
]
dependencies = [
    "matplotlib",
    "numba",
    "numpy",
    "biopython",
    "scipy",
    "pyhmmer; sys_platform != 'win32'",
]

[project.optional-dependencies]
unix = ["pyhmmer"]

[tool.setuptools.packages.find]
where = ["."] # list of folders that contain the packages (["."] by default)
include = [
    "dca*",
] # package names should match these glob patterns (["*"] by default)
namespaces = false # to disable scanning PEP 420 namespaces (true by default)

[project.urls]
Homepage = "https://github.com/utdal/py-mfdca/"
